+++ src/by/it/dorostchenok/jd02_threads/Main.java
@@ -1,4 +1,4 @@
-package by.it.dorostchenok.jd02_01;
+package by.it.dorostchenok.jd02_threads;
 
 public class Main {
     public static void main(String[] args) throws InterruptedException {
