+++ src/by/it/dorostchenok/jd02_threads/implementation/Market.java
@@ -1,4 +1,4 @@
-package by.it.dorostchenok.jd02_01.implementation;
+package by.it.dorostchenok.jd02_threads.implementation;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
